Blarney.BV

Description

This module represents the core of the Blarney HDL, upon which all hardware description features are built. It provides untyped bit vectors, i.e. bit vectors whose width is not specified in the type, and a set of associated circuit primitives. Hardware developers should not use these untyped primitives directly (unless they know what they are doing), because width-mistmatches are not checked. Any bit-vector can be evaluated symbolically, using a feature similar to Observable Sharing [1], to produce a netlist, i.e. a graph whose nodes are primitive component instances and whose edges are connections.

K. Claessen, D. Sands. Observable Sharing for Functional Circuit Description, ASIAN 1999.

Primitive component types

type InstId = Int

Every instance of a component in the circuit has a unique id

type OutputNumber = Int

Each output from a primitive component is numbered

type Width = Int

Bit vector width

type InputWidth = Int

Width of an input to a primitive

type OutputWidth = Int

Width of an output from a primitive

type BitIndex = Int

For indexing a bit vector

type RegFileId = Int

Identifier for register file primitive

data Prim

Primitive components

Constructors

Const OutputWidth Integer
Constant value (0 inputs, 1 output)

ConstBits OutputWidth ConstBit
Vector of constant bits

Add OutputWidth
Adder (2 inputs, 1 output)

Sub OutputWidth
Subtractor (2 inputs, 1 output)

Mul OutputWidth
Multiplier (2 inputs, 1 output)

Div OutputWidth
Quotient (2 inputs, 1 output)

Mod OutputWidth
Remainder (2 inputs, 1 output)

Not OutputWidth
Bitwise not (1 input, 1 output)

And OutputWidth
Bitwise and (2 inputs, 1 output)

Or OutputWidth
Bitwise or (2 inputs, 1 output)

Xor OutputWidth
Bitwise xor (2 inputs, 1 output)

ShiftLeft OutputWidth
Left shift (2 inputs, 1 output)

ShiftRight OutputWidth
Right shift (2 inputs, 1 output)

Equal InputWidth
Equality comparator (2 inputs, 1 bit output)

NotEqual InputWidth
Disequality comparator (2 inputs, 1 bit output)

LessThan InputWidth
Less-than comparator (2 inputs, 1 bit output)

LessThanEq InputWidth
Less-than-or-equal comparator (2 inputs, 1 bit output)

ReplicateBit OutputWidth
Replicate a bit (1 input, 1 output)

ZeroExtend InputWidth OutputWidth
Zero-extend a bit vector (1 input, 1 output)

SignExtend InputWidth OutputWidth
Sign-extend a bit vector (1 input, 1 output)

SelectBits InputWidth BitIndex BitIndex
Bit selection (compile-time range, 1 input, 1 output)

Concat InputWidth InputWidth
Bit vector concatenation (2 inputs, 1 output)

Mux OutputWidth
Multiplexer (3 inputs, 1 output)

CountOnes OutputWidth
Population count (1 input, 1 output)

Identity OutputWidth
Identity function (1 input, 1 output)

Register Integer InputWidth
Register with initial value (1 input, 1 output)

RegisterEn Integer InputWidth
Register with initial value and enable wire (2 inputs, 1 output)

BRAM
Single-port block RAM

Fields

ramInitFile :: Maybe String

ramAddrWidth :: Width

ramDataWidth :: Width

TrueDualBRAM
True dual-port block RAM

Fields

ramInitFile :: Maybe String

ramAddrWidth :: Width

ramDataWidth :: Width

Custom String [String] [(String, Int)] [Param]
Custom component (component name, input names, output names/widths, parameters)

Input OutputWidth String
External input (0 inputs, 1 output)

Output InputWidth String
External output (only used in RTL context, not expression context)

Display [DisplayArg]
Simulation-time I/O (only used in RTL context, not expression context)

Finish
Finish simulation (input: 1-bit enable wire)

RegFileMake String InputWidth InputWidth RegFileId
Register file declaration (only used in RTL context, not expression context)

RegFileRead OutputWidth RegFileId
Register file lookup (input: index, output: data)

RegFileWrite InputWidth InputWidth RegFileId
Register file update (inputs: write-enable, address, data)

data ConstBit

A constant bit is a 0, 1, or don't care

Constructors

Zero

One

DontCare

data DisplayArg

For the Display primitive: display a string literal or a bit-vector value of a given width

Constructors

DisplayArgString String

DisplayArgBit InputWidth

data Param

Custom components may have compile-time parameters. A parameter has a name and a value, both represented as strings

Constructors

String :-> String

lookupParam :: [Param] -> String -> String

Given a parameter name, return the parameter value

Untyped bit vectors

data BV

An untyped bit vector output wire from a primitive component instance

Constructors

BV

Fields

bvPrim :: Prim

What kind of primitive produced this bit vector?

bvInstRef :: IORef (Maybe InstId)

Unique id of primitive instance

bvInputs :: [BV]

Inputs to the primitive instance

bvOutNum :: OutputNumber

Output pin number

bvWidth :: Width

Width of this output pin

bvWidths :: [Width]

Width of each output pin

makePrim :: Prim -> [BV] -> [Int] -> [BV]

Helper function for creating an instance of a primitive component

makePrim1 :: Prim -> [BV] -> Int -> BV

Create instance of primitive component which has one output

Netlists

data Net

Netlists are lists of nets

Constructors

Net

Fields

netPrim :: Prim

netInstId :: InstId

netInputs :: [WireId]

netOutputWidths :: [Width]

type WireId = (InstId, OutputNumber)

A wire is uniquely identified by an instance id and an output number

newtype Flatten a

A reader/writer monad for accumulating the netlist

Constructors

Flatten

Fields

runFlatten :: FlattenR -> IO (FlattenW, a)

freshId :: Flatten Int

Obtain a fresh instance id

addNet :: Net -> Flatten ()

Add a net to the netlist

flatten :: BV -> Flatten WireId

Flatten bit vector to netlist

Bit-vector primitives

constBV :: Width -> Integer -> BV

Constant bit vector of given width

constBitsBV :: Width -> ConstBit -> BV

Bit vector of constant bits

addBV :: BV -> BV -> BV

Adder

subBV :: BV -> BV -> BV

Subtractor

mulBV :: BV -> BV -> BV

Multiplier

divBV :: BV -> BV -> BV

Quotient

modBV :: BV -> BV -> BV

Remainder

invBV :: BV -> BV

Bitwise not

andBV :: BV -> BV -> BV

Bitwise and

orBV :: BV -> BV -> BV

Bitwise or

xorBV :: BV -> BV -> BV

Bitwise xor

leftBV :: BV -> BV -> BV

Left shift

rightBV :: BV -> BV -> BV

Right shift

equalBV :: BV -> BV -> BV

Equality comparator

notEqualBV :: BV -> BV -> BV

Disequality comparator

lessThanBV :: BV -> BV -> BV

Less-than comparator

lessThanEqBV :: BV -> BV -> BV

Less-than-or-equal comparator

replicateBV :: Int -> BV -> BV

Replicate a bit

zeroExtendBV :: OutputWidth -> BV -> BV

Zero-extend a bit vector

signExtendBV :: OutputWidth -> BV -> BV

Sign-extend a bit vector

selectBV :: (BitIndex, BitIndex) -> BV -> BV

Bit selection

concatBV :: BV -> BV -> BV

Bit vector concatenation

muxBV :: BV -> (BV, BV) -> BV

Multiplexer

countOnesBV :: OutputWidth -> BV -> BV

Population count

idBV :: BV -> BV

Identity function

inputPinBV :: Width -> String -> BV

Input pin (named Verilog pin)

regBV :: Width -> BV -> BV -> BV

Register of given width with initial value

regEnBV :: Width -> BV -> BV -> BV -> BV

Register of given width with initial value and enable wire

ramBV :: OutputWidth -> Maybe String -> (BV, BV, BV) -> BV

Single-port block RAM. Input: one triple (address, data, write-enable)

dualRamBV :: OutputWidth -> Maybe String -> (BV, BV, BV) -> (BV, BV, BV) -> (BV, BV)

True dual-port block RAM. Input: two triples (address, data, write-enable)

regFileReadBV :: RegFileId -> OutputWidth -> BV -> BV

Read from register file

getInitBV :: BV -> Integer

Get the value of a constant bit vector, which may involve bit manipulations. Used to determine the initial value of a register.