Blarney.Bit

Description

This module provides size-typed bit vectors and circuit primitives, on top of Blarney's untyped bit vectors and circuit primitives. Hardware developers should always use the typed versions!

Typed bit-vectors

newtype Bit (n :: Nat)

Phantom type wrapping an untyped bit vector, capturing the bit-vector width. All bit vectors are members of the Num and Cmp classes.

Constructors

FromBV

Fields

toBV :: BV

widthOf :: KnownNat n => Bit n -> Int

Determine width of bit-vector from type

constant :: KnownNat n => Integer -> Bit n

Constant bit-vector

Bit-vector arithmetic

(.+.) :: Bit n -> Bit n -> Bit n infixl 6

Adder

(.-.) :: Bit n -> Bit n -> Bit n infixl 6

Subtractor

(.*.) :: Bit n -> Bit n -> Bit n infixl 7

Multiplier

(./.) :: Bit n -> Bit n -> Bit n infixl 7

Quotient

(.%.) :: Bit n -> Bit n -> Bit n infixl 7

Remainder

Bitwise operations on bit-vectors

inv :: Bit n -> Bit n

Bitwise invert

(.&.) :: Bit n -> Bit n -> Bit n infixl 7

Bitwise and

(.|.) :: Bit n -> Bit n -> Bit n infixl 5

Bitwise and

(.^.) :: Bit n -> Bit n -> Bit n infixl 6

Bitwise xor

(.<<.) :: Bit n -> Bit n -> Bit n

Shift left

(.>>.) :: Bit n -> Bit n -> Bit n

Shift right

Bit-vector comparison primitives

class Cmp a where

Methods

(.<.) :: a -> a -> Bit 1 infix 4

(.<=.) :: a -> a -> Bit 1 infix 4

(.==.) :: a -> a -> Bit 1 infix 4

(.>.) :: a -> a -> Bit 1 infix 4

(.>=.) :: a -> a -> Bit 1 infix 4

(.!=.) :: a -> a -> Bit 1 infix 4

Bit-vector width adjustment

rep :: KnownNat n => Bit 1 -> Bit n

Replicate bit

zeroExtend :: (KnownNat m, n <= m) => Bit n -> Bit m

Zero extension

signExtend :: (KnownNat m, n <= m) => Bit n -> Bit m

Sign extension

(#) :: Bit n -> Bit m -> Bit (n + m)

Bit-vector concatenation

upper :: (KnownNat m, m <= n) => Bit n -> Bit m

Extract most significant bits

lower :: (KnownNat m, m <= n) => Bit n -> Bit m

Extract least significant bits

split :: KnownNat n => Bit (n + m) -> (Bit n, Bit m)

Split bit vector

Bit-vector selection primitives

bits :: KnownNat m => (Int, Int) -> Bit n -> Bit m

Dynamically-typed bit selection

unsafeBits :: (Int, Int) -> Bit n -> Bit m

Untyped bit selection (try to avoid!)

range :: forall (hi :: Nat) (lo :: Nat) i o. (KnownNat hi, KnownNat lo, (lo + o) ~ (hi + 1), (hi + 1) <= i, o <= i) => Bit i -> Bit o

Statically-typed bit selection. Use type application to specify upper and lower indices.

bit :: Int -> Bit n -> Bit 1

Dynamically-typed bit indexing

index :: forall (i :: Nat) n. (KnownNat i, (i + 1) <= n) => Bit n -> Bit 1

Statically-typed bit indexing. Use type application to specify index.

Bit-vector registers

reg :: Bit n -> Bit n -> Bit n

Register

regEn :: Bit n -> Bit 1 -> Bit n -> Bit n

Register with enable wire

Misc. bit-vector operations

mux :: Bit 1 -> (Bit n, Bit n) -> Bit n

Multiplexer

countOnes :: Bit n -> Bit (Log2 n + 1)

Population count

liftNat :: Int -> (forall n. KnownNat n => Proxy n -> a) -> a

Lift integer value to type-level natural

bitListToBitVec :: KnownNat n => [Bit 1] -> Bit n

Convert list of bits to bit vector