Blarney.BitScan

Description

Dynamically typed pattern matching on bit-strings, similar to scanf. Here's an instruction decoder for a tiny subset of RISC-V, written using the match and ==> combinators provided by the module:

-- Semantics of add instruction
add :: Bit 5 -> Bit 5 -> Bit 5 -> RTL ()
add rs2 rs1 rd = display "add " rd ", " rs1 ", " rs2

-- Semantics of addi instruction
addi :: Bit 12 -> Bit 5 -> Bit 5 -> RTL ()
addi imm rs1 rd = display "addi " rd ", " rs1 ", " imm

-- Semantics of store instruciton
sw :: Bit 12 -> Bit 5 -> Bit 5 -> RTL ()
sw imm rs2 rs1 = display "sw " rs2 ", " rs1 "[" imm "]"

top :: RTL ()
top = do
  let instr :: Bit 32 = 0b1000000_00001_00010_010_00001_0100011

  match instr
    [
      "0000000   rs2[4:0]  rs1[4:0] 000 rd[4:0]  0110011" ==> add,
      "          imm[11:0] rs1[4:0] 000 rd[4:0]  0010011" ==> addi,
      "imm[11:5] rs2[4:0]  rs1[4:0] 010 imm[4:0] 0100011" ==> sw
    ]

  finish

Documentation

(==>) :: (KnownNat n, RHS rhs) => String -> rhs -> Bit n -> RTL () infix 7

Case alternative

match :: KnownNat n => Bit n -> [Bit n -> RTL ()] -> RTL ()

Case statement, with a subject and a list of alternatives