Blarney.Interface

Description

Rather than flattening the entire design hierarchy down to a single netlist, we want to maintain the modular structure when generating Verilog. This module allows Blarney functions to be turned into Verilog modules, and Verilog modules to be instantiated in a Blarney description.

Documentation

data Ifc a

The Ifc monad

class Interface a where

Any type in this class can be used as an input/output to a module when doing separate compilation.

Minimal complete definition

Nothing

Methods

writePort :: String -> a -> Ifc ()

readPort :: String -> Ifc a

writePort :: (Generic a, GInterface (Rep a)) => String -> a -> Ifc ()

readPort :: (Generic a, GInterface (Rep a)) => String -> Ifc a

class Modular a where

Any type in this class can be turned into a module when doing separate compilation.

Methods

makeMod :: a -> Module ()

makeInst :: String -> [Param] -> a

makeModule :: Modular a => a -> Module ()

makeInstance :: Modular a => String -> a

makeInstanceWithParams :: Modular a => String -> [Param] -> a