Blarney.Module

Description

We split the RTL monad into a Module monad and an Action monad, giving a more familar HDL structure in which modules instantiate other modules, and express behaviour through always blocks containing actions. Actions cannot instantiate modules.

Modules and actions

newtype Module a

A module is just a wrapper around the RTL monad

Constructors

M

Fields

runModule :: RTL a

newtype Action a

An action is just a wrapper around the RTL monad

Constructors

A

Fields

runAction :: RTL a

Lift actions to modules

always :: Action a -> Module a

Execute an action on every clock cycle

Action of doing nothing

noAction :: Action ()

Action of doing nothing

Conditional actions

when :: Bit 1 -> Action () -> Action ()

Conditional block over actions

whenR :: Bit 1 -> Action a -> Action a

Conditional block over actions with return value

switch :: Bits a => a -> [(a, Action ())] -> Action ()

Switch statement over actions

(-->) :: a -> b -> (a, b) infixl 0

Operator for switch statement alternatives

Validity of a value

class Valid t where

Validity of a value

Methods

valid :: t -> Bit 1

Variable value (read)

class Val v a | v -> a where

Variable value (read)

Methods

val :: Bits a => v -> a

Variable assignment (write)

class Assign v where

Variable assignment (write)

Methods

(<==) :: Bits a => v a -> a -> Action () infix 1

Registers

data Reg a

Register variables

Constructors

Reg

Fields

regId :: VarId

Unique identifier

regVal :: a

Current register value

makeReg :: Bits a => a -> Module (Reg a)

Create register with initial value

makeRegU :: Bits a => Module (Reg a)

Create wire with don't care initial value

makeDReg :: Bits a => a -> Module (Reg a)

A DReg holds the assigned value only for one cycle. At all other times, it has the given default value.

Wires

data Wire a

Wire variables

Constructors

Wire

Fields

wireId :: VarId

Unique identifier

wireVal :: a

Current wire value

active :: Bit 1

Is wire being assigned on this cycle?

makeWire :: Bits a => a -> Module (Wire a)

Create wire with default value

makeWireU :: Bits a => Module (Wire a)

Create wire with don't care default value

Read-Write and Write-Only interfaces

data ReadWrite a

Read-Write interface

Constructors

ReadWrite

Fields

rwReadVal :: a

rwWriteVal :: a -> Action ()

data WriteOnly a

Write-Only interface

Constructors

WriteOnly

Fields

woWriteVal :: a -> Action ()

Register files

data RegFile a d

Constructors

RegFile

Fields

(!) :: a -> d

update :: a -> d -> Action ()

makeRegFileInit :: forall a d. (Bits a, Bits d) => String -> Module (RegFile a d)

Create register file with initial contents

makeRegFile :: forall a d. (Bits a, Bits d) => Module (RegFile a d)

Create uninitialised register file

Other actions

finish :: Action ()

Terminate simulator

display :: Displayable a => a

Display statement

display_ :: Displayable a => a

Display statement (without new line)

External inputs and outputs

input :: KnownNat n => String -> Module (Bit n)

External input declaration

inputBV :: String -> Width -> Module BV

External input declaration (untyped)

output :: String -> Bit n -> Module ()

External output declaration

outputBV :: String -> BV -> Module ()

External output declaration (untyped)

Convert RTL to a netlist

addRoots :: [BV] -> Module ()

Add roots

netlist :: Module () -> IO [Net]

Convert module to a netlist