Blarney.Queue

Description

Documentation

data Queue a

Queue interface

Constructors

Queue

Fields

notEmpty :: Bit 1

notFull :: Bit 1

enq :: a -> Action ()

deq :: Action ()

canDeq :: Bit 1

first :: a

makeQueue :: Bits a => Module (Queue a)

A full-throughput 2-element queue implemented using 2 registers:

No combinatorial paths between sides.

There's a mux on the enqueue path.

makeSizedQueue :: Bits a => Int -> Module (Queue a)

A full-throughput N-element queue implemented using 2 registers and a RAM:

No combinatorial paths between sides.

There's a mux on the enqueue path.

makeSizedQueueCore :: Bits a => Int -> Module (Queue a)

This one has no output buffer (low latency, but not great for Fmax)

data ShiftQueueMode

There are modes of operation for the shift queue (below):

Optimise throughput: full throughput, but there's a combinatorial path between notFull and deq

Optimise Fmax: no combinatorial paths between sides, but max throughput = N/(N+1), where N is the queue capacity

Constructors

OptFmax

OptThroughput

makeShiftQueue :: Bits a => Int -> Module (Queue a)

An N-element queue implemented using a shift register:

No muxes: input element goes straight to a register and output element comes straight from a register.

N-cycle latency between enqueuing an element and being able to dequeue it, where N is the queue capacity.

This version optimised for Fmax.

makePipelineQueue :: Bits a => Int -> Module (Queue a)

This version is optimised for throughput

makeShiftQueueCore :: Bits a => ShiftQueueMode -> Int -> Module (Queue a)