Blarney.RAM

Description

Block RAM primitives

ram :: (Bits a, Bits d) => (a, d, Bit 1) -> d

Uninitialised block RAM. Reads new data on write.

ramInit :: (Bits a, Bits d) => String -> (a, d, Bit 1) -> d

Initilaised block RAM (contents taken from hex file). Reads new data on write.

ramTrueDual :: (Bits a, Bits d) => (a, d, Bit 1) -> (a, d, Bit 1) -> (d, d)

Uninitialised true dual-port block RAM. Reads new data on write. When read-address == write-address on different ports, reads old data.

ramTrueDualInit :: (Bits a, Bits d) => String -> (a, d, Bit 1) -> (a, d, Bit 1) -> (d, d)

RTL block RAM interface

data RAM a d

RAM interface

Constructors

RAM

Fields

load :: a -> Action ()

Issue load request

store :: a -> d -> Action ()

Issue store request

out :: d

Data out

writeEn :: Bit 1

Is there a store currently happening?

makeRAM :: (Bits a, Bits d) => Module (RAM a d)

Create uninitialised block RAM

makeRAMInit :: (Bits a, Bits d) => String -> Module (RAM a d)

Create block RAM with initial contents from hex file

makeTrueDualRAM :: (Bits a, Bits d) => Module (RAM a d, RAM a d)

Create true dual-port block RAM. When read-address == write-address on different ports, read old data.

makeTrueDualRAMInit :: (Bits a, Bits d) => String -> Module (RAM a d, RAM a d)

Create true dual-port block RAM with initial contents from hex file. When read-address == write-address on different ports, read old data.

makeDualRAM :: (Bits a, Bits d) => Module (RAM a d)

Create uninitialised dual-port RAM. One port used for reading and the other for writing.

makeDualRAMInit :: (Bits a, Bits d) => String -> Module (RAM a d)

Create dual-port RAM with initial contents from hex file. One port used for reading and the other for writing.

makeDualRAMForward :: (Bits a, Bits d) => Int -> Module (RAM a d)

Uninitialised dual-port forwarding block RAM. Read and write to same address yields new data.

makeDualRAMForwardInit :: (Bits a, Bits d) => Int -> String -> Module (RAM a d)

Dual-port forwarding block RAM with initial contents from hex file. Read and write to same address yields new data.