Blarney.RTL

Description

The module defines the RTL monad, supporting:

Mutable wires, registers and register files.

Conditional statements.

Simulation-time I/O.

Module input and output declarations.

RTL monad

data RTL a

The RTL monad, for register-transfer-level descriptions, is a fairly standard reader-writer-state monad.

Conditional statements

when :: Bit 1 -> RTL () -> RTL ()

RTL conditional block

whenR :: Bit 1 -> RTL a -> RTL a

RTL conditional block with return value

ifThenElseRTL :: Bit 1 -> RTL () -> RTL () -> RTL ()

If-then-else statement for RTL

switch :: Bits a => a -> [(a, RTL ())] -> RTL ()

RTL switch statement

(-->) :: a -> RTL () -> (a, RTL ()) infixl 0

Operator for switch statement alternatives

Mutable variables: registers and wires

class Var v where

Mutable variables

Methods

val :: Bits a => v a -> a

(<==) :: Bits a => v a -> a -> RTL () infix 1

data Reg a

Register variables

data Wire a

Wire variables

makeReg :: Bits a => a -> RTL (Reg a)

Create register with initial value

makeRegU :: Bits a => RTL (Reg a)

Create wire with don't care initial value

makeWire :: Bits a => a -> RTL (Wire a)

Create wire with default value

makeWireU :: Bits a => RTL (Wire a)

Create wire with don't care default value

makeDReg :: Bits a => a -> RTL (Reg a)

A DReg holds the assigned value only for one cycle. At all other times, it has the given default value.

Simulation-time statements

class Displayable a

To support a display statement with variable number of arguments

Minimal complete definition

disp

display :: Displayable a => a

Display statement

finish :: RTL ()

Terminate simulator

External inputs and outputs

input :: KnownNat n => String -> RTL (Bit n)

RTL external input declaration

output :: String -> Bit n -> RTL ()

RTL external output declaration

inputBV :: String -> Width -> RTL BV

RTL external input declaration (untyped)

outputBV :: String -> BV -> RTL ()

RTL external output declaration (untyped)

Register files

data RegFile a d

Register file interface

Constructors

RegFile

Fields

(!) :: a -> d

update :: a -> d -> RTL ()

makeRegFileInit :: forall a d. (Bits a, Bits d) => String -> RTL (RegFile a d)

Create register file with initial contents

makeRegFile :: forall a d. (Bits a, Bits d) => RTL (RegFile a d)

Create uninitialised register file

Convert RTL to a netlist

netlist :: RTL () -> IO [Net]

Convert RTL monad to a netlist