Blarney.Recipe

Description

Recipe is a lightweight imperative language, aidingconcise definitions of complex state machines.

Documentation

data Recipe where

Abstract syntax of Recipe

Constructors

Tick :: Recipe

(:=) :: (Var v, Bits a) => v a -> a -> Recipe infixl 1

Wait :: Bit 1 -> Recipe

When :: Bit 1 -> RTL () -> Recipe

Do :: [RTL ()] -> Recipe

RTL :: RTL () -> Recipe

Seq :: [Recipe] -> Recipe

Par :: [Recipe] -> Recipe

If :: Bit 1 -> Recipe -> Recipe

While :: Bit 1 -> Recipe -> Recipe

run :: Bit 1 -> Recipe -> RTL (Bit 1)

Run a recipe. Take a go pulse and return a finish pulse.

runOnce :: Recipe -> RTL ()

Run a recipe with a start pulse that is high only on the first cycle of execution, and ignore the finish pulse.